darwin: separate basic iokit routines from event support
