+++ osquery/core/CMakeLists.txt
@@ -13,6 +13,7 @@ if(APPLE)
   set(OSQUERY_CORE_PLATFORM
     ${OSQUERY_CORE_PLATFORM}
     darwin/conversions.cpp
+    darwin/iokit.cpp
   )
 endif()
