+++ osquery/events/darwin/iokit.cpp
@@ -18,12 +18,6 @@
 
 namespace osquery {
 
-const std::string kIOUSBDeviceClassName_ = "IOUSBDevice";
-const std::string kIOPCIDeviceClassName_ = "IOPCIDevice";
-const std::string kIOPlatformExpertDeviceClassName_ = "IOPlatformExpertDevice";
-const std::string kIOACPIPlatformDeviceClassName_ = "IOACPIPlatformDevice";
-const std::string kIOPlatformDeviceClassname_ = "IOPlatformDevice";
-
 REGISTER(IOKitEventPublisher, "event_publisher", "iokit");
 
 struct DeviceTracker : private boost::noncopyable {
@@ -35,79 +29,6 @@ struct DeviceTracker : private boost::noncopyable {
   io_object_t notification{0};
 };
 
-IOKitPCIProperties::IOKitPCIProperties(const std::string& compatible) {
-  auto properties = osquery::split(compatible, " ");
-  if (properties.size() < 2) {
-    return;
-  }
-
-  size_t prop_index = 0;
-  if (properties[1].find("pci") == 0 && properties[1].find("pciclass") != 0) {
-    // There are two sets of PCI definitions.
-    prop_index = 1;
-  } else if (properties[0].find("pci") != 0) {
-    return;
-  }
-
-  auto vendor = osquery::split(properties[prop_index++], ",");
-  vendor_id = vendor[0].substr(3);
-  model_id = (vendor[1].size() == 3) ? "0" + vendor[1] : vendor[1];
-
-  if (properties[prop_index].find("pciclass") == 0) {
-    // There is a class definition.
-    pci_class = properties[prop_index++].substr(9);
-  }
-
-  if (properties.size() > prop_index) {
-    // There is a driver/ID.
-    driver = properties[prop_index];
-  }
-}
-
-std::string getIOKitProperty(const CFMutableDictionaryRef& details,
-                             const std::string& key) {
-  std::string value;
-
-  // Get a property from the device.
-  auto cfkey = CFStringCreateWithCString(
-      kCFAllocatorDefault, key.c_str(), kCFStringEncodingUTF8);
-  auto property = CFDictionaryGetValue(details, cfkey);
-  CFRelease(cfkey);
-
-  // Several supported ways of parsing IOKit-encoded data.
-  if (property) {
-    if (CFGetTypeID(property) == CFNumberGetTypeID()) {
-      value = stringFromCFNumber((CFDataRef)property);
-    } else if (CFGetTypeID(property) == CFStringGetTypeID()) {
-      value = stringFromCFString((CFStringRef)property);
-    } else if (CFGetTypeID(property) == CFDataGetTypeID()) {
-      value = stringFromCFData((CFDataRef)property);
-    } else if (CFGetTypeID(property) == CFBooleanGetTypeID()) {
-      value = (CFBooleanGetValue((CFBooleanRef)property)) ? "1" : "0";
-    }
-  }
-
-  return value;
-}
-
-long long int getNumIOKitProperty(const CFMutableDictionaryRef& details,
-                                  const std::string& key) {
-  // Get a property from the device.
-  auto cfkey = CFStringCreateWithCString(
-      kCFAllocatorDefault, key.c_str(), kCFStringEncodingUTF8);
-  auto property = CFDictionaryGetValue(details, cfkey);
-  CFRelease(cfkey);
-
-  // Several supported ways of parsing IOKit-encoded data.
-  if (property && CFGetTypeID(property) == CFNumberGetTypeID()) {
-    CFNumberType type = CFNumberGetType((CFNumberRef)property);
-    long long int value;
-    CFNumberGetValue((CFNumberRef)property, type, &value);
-    return value;
-  }
-
-  return 0;
-}
 
 void IOKitEventPublisher::restart() {
   static std::vector<const std::string*> device_classes = {
