+++ osquery/tables/system/darwin/iokit_registry.cpp
@@ -11,7 +11,7 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
-#include "osquery/events/darwin/iokit.h"
+#include "osquery/core/darwin/iokit.hpp"
 
 namespace osquery {
 namespace tables {
