+++ osquery/tables/system/darwin/kernel_info.cpp
@@ -15,7 +15,7 @@
 #include <osquery/logger.h>
 #include <osquery/tables.h>
 
-#include "osquery/events/darwin/iokit.h"
+#include "osquery/core/darwin/iokit.hpp"
 #include "osquery/tables/system/efi_misc.h"
 
 namespace osquery {
