+++ osquery/tables/system/darwin/sip_config.cpp
@@ -13,7 +13,7 @@
 #include <osquery/sql.h>
 #include <osquery/logger.h>
 
-#include "osquery/events/darwin/iokit.h"
+#include "osquery/core/darwin/iokit.hpp"
 
 namespace osquery {
 namespace tables {
