+++ osquery/tables/system/darwin/smbios_tables.cpp
@@ -18,7 +18,7 @@
 
 #include <osquery/tables.h>
 
-#include "osquery/events/darwin/iokit.h"
+#include "osquery/core/darwin/iokit.hpp"
 #include "osquery/tables/system/smbios_utils.h"
 
 namespace osquery {
